Go back to single shared view controller between iOS/tvOS (for now)
